third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-step-into-async.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="resources/framework.js"></script>
<script>

var dummy = function FAIL_should_not_pause_here() { return 0; };

function testFunction()
{
    debugger; // <- will turn on async call stacks here.

    setTimeout(dummy, 0);
    setTimeout(dummy.bind(null), 10);

    setTimeout(callback1, 20); // <- StepIntoAsync

    setTimeout(dummy, 0);
    setTimeout(dummy.bind(null), 10);
}

function callback1()
{
    var func = setTimeout.bind(null, callback2, 0);
    func = setTimeout.bind(null, func, 0);
    func = setTimeout.bind(null, func, 0);
    setTimeout(func); // <- StepIntoAsync
    Promise.resolve().then(dummy).then(dummy).then(dummy);
    return 1;
}

function callback2()
{
    var func = Framework.willSchedule(callback3);
    func = Framework.willSchedule(func);
    func = Framework.willSchedule(func);
    Framework.schedule(func); // <- StepIntoAsync
    Promise.resolve().then(dummy).then(dummy).then(dummy);
    setTimeout(dummy, 0);
    return 2;
}

function callback3()
{
    var func = Framework.bind(Framework.safeRun, null, Framework.createButtonWithEventListenersAndClick, Framework.throwFrameworkException, callback4);
    Framework.schedule(func); // <- StepIntoAsync
    return 3;
}

function callback4()
{
    var func = Framework.bind(Framework.createButtonWithEventListenersAndClick, null, callback5);
    Framework.schedule(func); // <- StepIntoAsync
    return 4;
}

function callback5()
{
    var resolve;
    var promise = new Promise(function(a, b) { resolve = a; });
    promise
        .then(JSON.stringify)
        .then(JSON.parse)
        .then(Framework.empty)
        .then(Framework.createButtonWithEventListenersAndClick)
        .then(callback6)
        .then(dummy)
        .catch(dummy);
    resolve({foo: 42}); // <- StepIntoAsync
    setTimeout(dummy, 0);
    return 5;
}

function callback6()
{
    Promise.resolve({foo: 43}) // <- StepIntoAsync
        .then(JSON.stringify)
        .then(JSON.parse)
        .then(Framework.empty)
        .then(Framework.createButtonWithEventListenersAndClick)
        .then(callback7)
        .then(dummy)
        .catch(dummy);
    setTimeout(dummy, 0);
    return 6;
}

function callback7()
{
    Promise.resolve(42)
        .then(
            function inner()
            {
                debugger;
                return Promise.reject(new Error("my error")); // <- StepIntoAsync
            }
        )
        .then(dummy)
        .catch(Framework.throwFrameworkException)
        .catch(callback8);
    return 7;
}

function callback8()
{
    var obj = {};
    Object.observe(obj, callback9);
    debugger;
    obj.foo = 1; // <- StepIntoAsync
    setTimeout(dummy);
    return 8;
}

function callback9()
{
    var iframe = document.getElementById("iframe");
    var win = iframe.contentWindow;
    win.postMessage("skip debugger", "*"); // <- StepIntoAsync
    setTimeout(dummy, 0);
    setTimeout(callback10, 10)
    setTimeout(dummy, 0);
    Promise.resolve().then(dummy);
    return 9;
}

function callback10()
{
    debugger;
    Promise.resolve({foo: 44}) // <- StepIntoAsync
        .then(
            function inner1()  // <- StepIntoAsync when in inner1 should lead debugger to the
            {                  // next promise handler in the chain, i.e. the one, which will
            }                  // receive the result returned by inner1, that is inner2.
        )
        .then(Framework.empty)
        .then(
            function inner2()
            {
            }
        )
        .then(dummy);
    setTimeout(dummy, 0);
    return 10;
}

function test()
{
    var maxAsyncCallStackDepth = 4;
    var frameworkRegexString = "/framework\\.js$";
    WebInspector.settings.skipStackFramesPattern.set(frameworkRegexString);

    InspectorTest.startDebuggerTest(step1, true);

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2()
    {
        DebuggerAgent.setAsyncCallStackDepth(maxAsyncCallStackDepth, step3);
    }

    function step3()
    {
        var actions = [
            "StepOver", "StepOver", "StepOver", "Print", // on setTimeout(callback1)
            "StepIntoAsync", "Print", // at callback1
            "StepIntoAsync", "Print", // at callback2
            "StepIntoAsync", "Print", // at callback3
            "StepIntoAsync", "Print", // at callback4
            "StepIntoAsync", "Print", // at callback5
            "StepIntoAsync", "Print", // at callback6
            "StepIntoAsync", "Print", // at callback7
            "Resume", // now paused at debugger in inner()
            "StepIntoAsync", "Print", // at callback8
            "Resume",
            "StepIntoAsync", "Print", // at callback9
            "StepIntoAsync", "Print", // at onmessage handler in iframe
            "Resume", // now paused at debugger in callback10
            "StepIntoAsync", "Print", // in inner1
            "StepIntoAsync", "Print", // in inner2
        ];
        InspectorTest.waitUntilPausedAndPerformSteppingActions(actions, step4);
    }

    function step4()
    {
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests debugger StepIntoAsync action.
</p>
<div><iframe src="resources/post-message-listener.html"
    id="iframe" width="800" height="100" style="border: 1px solid black;">
</iframe></div>
</body>
</html>
